Query refactoring to fix static-data
Query methods have been rewritten and are now in the clj-lolapi.query
namespace. They now take an optional params map, which is used to send
extra GET parameters to the server.

As a result, the static_data/champions and static_data/champion now
return *all* the data about the target champions (champData=all).